+++ tests/fixtures/model.py
@@ -10,6 +10,7 @@ from supadantic.models import BaseSBModel
 class ModelMock(BaseSBModel):
     id: int | None = None
     name: str
+    age: int | None = None
     some_optional_list: list[str] | None = None
     some_optional_tuple: tuple[str, ...] | None = None